requirements.txt
streamlit
openai
python-dotenv
fpdf
pytesseract
pillow
gitpython
